chore: split upgrade.go into 3 files
