+++ pkg/upgrade/util.go
@@ -1,0 +1,40 @@
+package upgrade
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"strings"
+
+	ofapi "github.com/operator-framework/api/pkg/operators/v1alpha1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+func GetOperatorNamespace() (string, error) {
+	data, err := os.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/namespace")
+	if err == nil {
+		if ns := strings.TrimSpace(string(data)); len(ns) > 0 {
+			return ns, nil
+		}
+	}
+
+	return "", err
+}
+
+// getClusterServiceVersion retries the clusterserviceversions available in the operator namespace.
+func getClusterServiceVersion(ctx context.Context, c client.Client, watchNameSpace string) (*ofapi.ClusterServiceVersion, error) {
+	clusterServiceVersionList := &ofapi.ClusterServiceVersionList{}
+	if err := c.List(ctx, clusterServiceVersionList, client.InNamespace(watchNameSpace)); err != nil {
+		return nil, fmt.Errorf("failed listign cluster service versions: %w", err)
+	}
+
+	for _, csv := range clusterServiceVersionList.Items {
+		for _, operatorCR := range csv.Spec.CustomResourceDefinitions.Owned {
+			if operatorCR.Kind == "DataScienceCluster" {
+				return &csv, nil
+			}
+		}
+	}
+
+	return nil, nil
+}
